+++ src/main.jsx
@@ -5,10 +5,10 @@ import { Auth0Provider } from "@auth0/auth0-react";
 import App from "./App.jsx";
 import "./index.css";
 import Categories from "./components/Categories.jsx";
-import LaptopsProductPage from "./components/LaptopsProductPage.jsx";
-import PhonesProductPage from "./components/PhonesProductPage.jsx";
-import AccessoriesProductPage from "./components/AccessoriesProductPage.jsx";
-import TabletsProductPage from "./components/TabletsProductPage.jsx";
+import LaptopsProductPage from "./components/productLists/LaptopsProductPage.jsx";
+import PhonesProductPage from "./components/productLists/PhonesProductPage.jsx";
+import AccessoriesProductPage from "./components/productLists/AccessoriesProductPage.jsx";
+import TabletsProductPage from "./components/productLists/TabletsProductPage.jsx";
 import AboutUs from "./components/AboutUs.jsx";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
